Try to add some sanity to dev process via "tests"
By writing a python script to copy the code and filter out
jenkins-specific stuff (at this point just the NonCPS annotations),
we make it possible to check the code.

If you wanted to be really test-driven you could write actual
tests in test.groovy; for now I'm just importing TaskTracker, which
is equivalent to making sure it compiles.